// src/generated/get-hotel-content.ts
export interface components {
    schemas: {
        GetHotelContentRequest: {
            GetHotelContentRQ: {
                POS?: components["schemas"]["POS"];
                SearchCriteria: components["schemas"]["SearchCriteria"];
            };
        };
        GetHotelContentResponse: {
            /** @description Contains property static and media details based on request parameters. */
            GetHotelContentRS?: {
                ApplicationResults?: components["schemas"]["ApplicationResults"];
                HotelContentInfos?: components["schemas"]["HotelContentInfos"];
            };
        };
// ...
        HotelRef: {
// ...
            CodeContext: "SABRE" | "GLOBAL";
        };
        /** @description Contains property and facilities description. */
        DescriptiveInfoRef: {
            /**
             * @description If true, property information is expected in the response.
             * @default false
             * @example true
             */
            PropertyInfo: boolean;
            /**
             * @description If true, location information is expected in the response.
             * @default false
             * @example true
             */
            LocationInfo: boolean;
// ...
        };
// ...
        MediaRef: {
            /**
             * @description Identifies maximum media items requested by the user. Valid values are "ALL" or 1 to 99.
             * @default 1
             * @example 10
             */
            MaxItems: string;
            MediaTypes?: components["schemas"]["MediaTypes"];
            Categories?: components["schemas"]["Categories"];
            RoomTypeCodes?: components["schemas"]["RoomTypeCodes"];
            AdditionalInfo?: components["schemas"]["AdditionalInfo"];
            Languages?: components["schemas"]["Languages"];
        };
        /** @description Contains types of media which are expected in the response. */
        MediaTypes: {
            Images?: components["schemas"]["Images"];
            PanoramicMedias?: components["schemas"]["PanoramicMedias"];
            Videos?: components["schemas"]["Videos"];
        };
        /** @description Contains details of the image types. */
        Images: {
            Image?: components["schemas"]["Image"];
        };
// ...
        PanoramicMedias: {
            PanoramicMedia?: components["schemas"]["PanoramicMedia"];
        };
// ...
        Videos: {
            Video?: components["schemas"]["Video"];
        };
// ...
        Category: {
            /**
             * @description The OTA value of category code for which media items needs to be received.
             * @example 1
             */
            Code: number;
        }[];
// ...
        AdditionalInfo: {
            Info: components["schemas"]["Info"];
        };
// ...
        ElementStructure: {
            /** @description Specifies the source of error when processing the request. */
            type?: string;
            /** @description Contains current application server time stamp. */
            timeStamp?: string;
            /** @description Lists the application response status. */
            SystemSpecificResults?: {
                /** @description Contains current application server time stamp. */
                timeStamp?: string;
                /** @description If present, provides the name of the field for which the results apply. */
                reference?: string;
                /** @description Contains an abbreviated version of the error in text format. */
                ShortText?: Record<string, never>;
                /** @description Contains the identity of the unknown or misspelled tag that caused the processing error. */
                Element?: Record<string, never>;
                /** @description Contains the identity of the record that failed among a group of records. This value may contain a concatenation of a unique failed transaction ID with specific record(s) associated with that transaction. */
                RecordID?: Record<string, never>;
                /** @description Contains an online description of the error that occurred. */
                DocURL?: Record<string, never>;
                /** @description Contains the host system common run to create the result. */
                HostCommand?: {
                    /** @description The Line Number Interchange Address Terminal Address (LNIATA) number. */
                    LNIATA?: string;
                };
                /** @description Lists application-specific codes and messages. A textual description to provide more information about the specific condition, warning, or error with the code attribute as a numeric value. */
                Message?: {
                    /**
                     * @description The code, if returned, of an error or warning.
                     * @example ERR.0381
                     */
                    code?: string;
                    /**
                     * @description The message associated to the above code.
                     * @example Invalid check-in date
                     */
                    value?: string;
                }[];
            }[];
        }[];
        /** @description Wrapper around HotelContentInfo. */
        HotelContentInfos: {
            HotelContentInfo: components["schemas"]["HotelContentInfo"];
        };
        /** @description parent object which contains all data for a single hotel in the response - static content, rates and media content */
        HotelContentInfo: {
            HotelInfo: components["schemas"]["HotelInfo"];
            HotelDescriptiveInfo?: components["schemas"]["HotelDescriptiveInfo"];
            HotelMediaInfo?: components["schemas"]["HotelMediaInfo"];
        };
// ...
        TierLabels: {
            TierLabel?: components["schemas"]["TierLabel"];
        };
// ...
        PropertyTypeInfo: {
            PropertyType?: components["schemas"]["PropertyType"];
        };
// ...
        Policies: {
            Policy: components["schemas"]["Policy"];
        };
// ...
        PropertyQualityInfo: {
            PropertyQuality?: components["schemas"]["PropertyQuality"];
        };
// ...
        LocationInfo: {
            /**
             * @description Specifies the geographical latitude of the hotel location.
             * @example 36.1548
             */
            Latitude?: number;
            /**
             * @description Specifies the geographical longitude of the hotel location.
             * @example -95.990356
             */
            Longitude?: number;
            Address?: components["schemas"]["Address"];
            Neighborhoods?: components["schemas"]["Neighborhoods"];
            Contact?: components["schemas"]["Contact"];
        };
// ...
        CityName: {
            /**
             * @description The City Code.
             * @example TUL
             */
            CityCode?: string;
            /**
             * @description The City Name.
             * @example Tulsa
             */
            value?: string;
        };
        StateProv: {
            /**
             * @description The property, state or province code.
             * @example O
             */
            StateCode?: string;
            /**
             * @description The state or province name.
             * @example Oklahoma
             */
            value?: string;
        };
// ...
        Neighborhoods: {
            Neighborhood?: components["schemas"]["Neighborhood"];
        };
// ...
        Amenities: {
            Amenity?: components["schemas"]["Amenity"];
        };
// ...
        SecurityFeatures: {
            SecurityFeature?: components["schemas"]["SecurityFeature"];
        };
// ...
        "HotelDescriptiveInfo.Description": {
            /** @description Contains the descriptive information returned in response. */
            Text?: {
                /**
                 * @description The type of Descriptive Information required in response.
                 * @example Dining
                 * @enum {string}
                 */
                Type: "ShortDescription" | "Dining" | "Facilities" | "Recreation" | "Services" | "Attractions" | "CancellationPolicy" | "DepositPolicy" | "Directions" | "Policies" | "SafetyInfo" | "TransportationInfo" | "GuaranteePolicy";
                /**
                 * @description Textual description for the type returned under 'Type'
                 * @example Bluefire Grill   Full service dining, Breakfast Buffet,    mediterrranian menu with midwest   specialties.     **Closed everyday from 1400-1700** - Bevende Lounge   Art Deco atmosphere. Snack menu available. - Bean Counter    Quick pickup and go, Starbuck*s coffee, danish,    donuts, bagels, etc. - Room Service   Available 0600-2300
                 */
                value?: string;
            };
        }[];
        /** @description Contains a list of media types, each of which represents a media type (Image / Panoramic Image / Video) available for a particular hotel code. */
        HotelMediaInfo: {
            MediaItems: components["schemas"]["MediaItems"];
        };
        MediaItems: {
            MediaItem?: components["schemas"]["MediaItem"];
        };
// ...
        "response.Image": {
            /**
             * Format: uri
             * @description The Image URL.
             * @example http://vcmp-hotels.cert.sabre.com/image/upload/t_vcmp_thumb/hotel/i/1/jdbkyzhiy4tmmm0a30e3.jpg
             */
            Url: string;
            /**
             * @description The types of images.
             * @example THUMBNAIL
             * @enum {string}
             */
            Type: "ORIGINAL" | "THUMBNAIL" | "SMALL" | "MEDIUM" | "LARGE";
            /**
             * @description The height of the image in pixels.
             * @example 100
             */
            Height: number;
            /**
             * @description The width of the image in pixels.
             * @example 150
             */
            Width: number;
        }[];
        /** @description Contains Panoramic Image items for property. */
        PanoramicMediasItems: {
            PanoramicMedia?: components["schemas"]["response.PanoramicMedia"];
        };
        "response.PanoramicMedia": {
            /**
             * Format: uri
             * @description The Panoramic Image's URL.
             * @example https://www.iceportal.com/brochures/Media/Show/photoThumb.aspx?did=8845&brochureid=6354&mtype=26162&publicid=22707462&instanceid=9&resizing=default
             */
            Url: string;
            /**
             * @description The Panoramic Image's types.
             * @example HD360
             * @enum {string}
             */
            Type: "HD360";
            /**
             * @description The Panoramic Image's height in pixels.
             * @example 100
             */
            Height: number;
            /**
             * @description The Panoramic Image's width in pixels.
             * @example 150
             */
            Width: number;
        }[];
        /** @description Contains video items for the given property. */
        VideosItems: {
            Video?: components["schemas"]["response.Video"];
        };
        "response.Video": {
            /**
             * Format: uri
             * @description The Video item 's URL.
             * @example http://s3.amazonaws.com/iceportaldata/brochures/20416/video021016165607956.mp4
             */
            Url: string;
            /**
             * @description The video item's types.
             * @example VIDEO360
             * @enum {string}
             */
            Type: "VIDEOTHUMBNAIL" | "VIDEO360" | "VIDEO720" | "ALL";
            /**
             * @description The video item's height.
             * @example 100
             */
            Height: number;
            /**
             * @description The video item's width.
             * @example 150
             */
            Width: number;
        }[];
        /** @description Contains the type of category to which the media item belongs. it is mapped to OTA code PCI. */
        "response.Category": {
            /**
             * @description The media item category code. It is mapped to the PCI OTA code.
             * @example 3
             */
            CategoryCode?: number;
            Description: components["schemas"]["response.Category.Description"];
        };
// ...
        "response.AdditionalInfo.Info": {
            /**
             * @description Specifies the type of additional info.
             * @example CAPTION
             * @enum {string}
             */
            Type: "CAPTION" | "ROOM_TYPE_CODE";
            Description: components["schemas"]["response.AdditionalInfo.Info.Description"];
        }[];
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}